+++ src/main/java/alfred/command/Command.java
@@ -1,3 +1,10 @@
+package alfred.command;
+
+import alfred.task.TaskList;
+import alfred.ui.Ui;
+import alfred.storage.Storage;
+import alfred.exceptions.AlfredException;
+
 public abstract class Command {
 
     // Maybe can add storage.write() after every execute command?
