+++ src/main/java/alfred/exceptions/AlfredException.java
@@ -1,3 +1,5 @@
+package alfred.exceptions;
+
 public class AlfredException extends Exception {
     public AlfredException(String message) {
         super("OOPS!!! " + message);
