+++ src/main/java/alfred/storage/Storage.java
@@ -1,8 +1,9 @@
+package alfred.storage;
+
+import alfred.task.*;
+import alfred.exceptions.AlfredException;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.sql.Array;
 import java.util.ArrayList;
 import java.util.Scanner;
