+++ src/main/java/alfred/task/Event.java
@@ -1,3 +1,5 @@
+package alfred.task;
+
 import java.time.LocalDateTime;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
