+++ src/main/java/alfred/task/Task.java
@@ -1,3 +1,5 @@
+package alfred.task;
+
 import java.time.LocalDate;
 
 public class Task {
