+++ src/main/java/alfred/task/ToDo.java
@@ -1,3 +1,4 @@
+package alfred.task;
 public class ToDo extends Task{
 
     public ToDo(String description) {
